dual core capable
